# gen/restapi/operations/user/profile_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package user

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"

	"github.com/wasimkhan042/ustore-auth/gen/models"
)

// ProfileOKCode is the HTTP code returned for type ProfileOK
const ProfileOKCode int = 200

/*ProfileOK Success response when item is added successfully

swagger:response profileOK
*/
type ProfileOK struct {

	/*
	  In: Body
	*/
	Payload *models.Profile `json:"body,omitempty"`
}

// NewProfileOK creates ProfileOK with default headers values
func NewProfileOK() *ProfileOK {

	return &ProfileOK{}
}

// WithPayload adds the payload to the profile o k response
func (o *ProfileOK) WithPayload(payload *models.Profile) *ProfileOK {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the profile o k response
func (o *ProfileOK) SetPayload(payload *models.Profile) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *ProfileOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(200)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}

// ProfileBadRequestCode is the HTTP code returned for type ProfileBadRequest
const ProfileBadRequestCode int = 400

/*ProfileBadRequest Bad Request

swagger:response profileBadRequest
*/
type ProfileBadRequest struct {
}

// NewProfileBadRequest creates ProfileBadRequest with default headers values
func NewProfileBadRequest() *ProfileBadRequest {

	return &ProfileBadRequest{}
}

// WriteResponse to the client
func (o *ProfileBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(400)
}

// ProfileNotFoundCode is the HTTP code returned for type ProfileNotFound
const ProfileNotFoundCode int = 404

/*ProfileNotFound User not found

swagger:response profileNotFound
*/
type ProfileNotFound struct {
}

// NewProfileNotFound creates ProfileNotFound with default headers values
func NewProfileNotFound() *ProfileNotFound {

	return &ProfileNotFound{}
}

// WriteResponse to the client
func (o *ProfileNotFound) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(404)
}

// ProfileInternalServerErrorCode is the HTTP code returned for type ProfileInternalServerError
const ProfileInternalServerErrorCode int = 500

/*ProfileInternalServerError Server error

swagger:response profileInternalServerError
*/
type ProfileInternalServerError struct {

	/*
	  In: Body
	*/
	Payload string `json:"body,omitempty"`
}

// NewProfileInternalServerError creates ProfileInternalServerError with default headers values
func NewProfileInternalServerError() *ProfileInternalServerError {

	return &ProfileInternalServerError{}
}

// WithPayload adds the payload to the profile internal server error response
func (o *ProfileInternalServerError) WithPayload(payload string) *ProfileInternalServerError {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the profile internal server error response
func (o *ProfileInternalServerError) SetPayload(payload string) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *ProfileInternalServerError) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(500)
	payload := o.Payload
	if err := producer.Produce(rw, payload); err != nil {
		panic(err) // let the recovery middleware deal with this
	}
}